+++ threedi_modelchecker/__init__.py
@@ -1,6 +1,4 @@
 from .model_checks import *  # NOQA
-from .threedi_database import *  # NOQA
-
 
 # fmt: off
 __version__ = '0.35.3.dev0'