+++ threedi_modelchecker/checks/base.py
@@ -1,15 +1,10 @@
-from abc import ABC
-from abc import abstractmethod
+from abc import ABC, abstractmethod
 from enum import IntEnum
-from geoalchemy2 import functions as geo_func
-from geoalchemy2.types import Geometry
-from sqlalchemy import and_
-from sqlalchemy import false
-from sqlalchemy import func
-from sqlalchemy import types
+from typing import List, NamedTuple
+
+from sqlalchemy import and_, false, func, types
 from sqlalchemy.orm.session import Session
-from typing import List
-from typing import NamedTuple
+from threedi_schema.domain import custom_types
 
 
 class CheckLevel(IntEnum):
@@ -257,7 +252,7 @@ def _sqlalchemy_to_sqlite_types(column_type):
         return ["integer"]
     elif isinstance(column_type, types.Date):
         return ["text"]
-    elif isinstance(column_type, Geometry):
+    elif isinstance(column_type, custom_types.Geometry):
         return ["blob"]
     elif isinstance(column_type, types.TIMESTAMP):
         return ["text"]
@@ -273,7 +268,7 @@ class GeometryCheck(BaseCheck):
     def get_invalid(self, session):
         q_invalid = self.to_check(session)
         invalid_geometries = q_invalid.filter(
-            geo_func.ST_IsValid(self.column) != True, self.column != None
+            func.ST_IsValid(self.column) != True, self.column != None
         )
         return invalid_geometries.all()
 
@@ -296,7 +291,7 @@ class GeometryTypeCheck(BaseCheck):
             # skip in case of generic GEOMETRY column
             return q_invalid.filter(false())
         invalid_geometry_types_q = q_invalid.filter(
-            geo_func.ST_GeometryType(self.column) != expected_geometry_type,
+            func.ST_GeometryType(self.column) != expected_geometry_type,
             self.column != None,
         )
         return invalid_geometry_types_q.all()