+++ threedi_modelchecker/checks/factories.py
@@ -1,12 +1,14 @@
-from ..threedi_model import custom_types
-from .base import EnumCheck
-from .base import ForeignKeyCheck
-from .base import GeometryCheck
-from .base import GeometryTypeCheck
-from .base import NotNullCheck
-from .base import TypeCheck
-from .base import UniqueCheck
-from geoalchemy2.types import Geometry
+from threedi_schema import custom_types
+
+from .base import (
+    EnumCheck,
+    ForeignKeyCheck,
+    GeometryCheck,
+    GeometryTypeCheck,
+    NotNullCheck,
+    TypeCheck,
+    UniqueCheck,
+)
 
 
 def get_level(table, column, level_map):
@@ -64,7 +66,7 @@ def generate_geometry_checks(table, custom_level_map=None, **kwargs):
     custom_level_map = custom_level_map or {}
     geometry_checks = []
     for column in table.columns:
-        if type(column.type) == Geometry:
+        if isinstance(column.type, custom_types.Geometry):
             level = get_level(table, column, custom_level_map)
             geometry_checks.append(GeometryCheck(column, level=level, **kwargs))
     return geometry_checks
@@ -74,7 +76,7 @@ def generate_geometry_type_checks(table, custom_level_map=None, **kwargs):
     custom_level_map = custom_level_map or {}
     geometry_type_checks = []
     for column in table.columns:
-        if type(column.type) == Geometry:
+        if isinstance(column.type, custom_types.Geometry):
             level = get_level(table, column, custom_level_map)
             geometry_type_checks.append(
                 GeometryTypeCheck(column, level=level, **kwargs)