+++ threedi_modelchecker/checks/other.py
@@ -1,19 +1,12 @@
-from ..threedi_model import constants
-from ..threedi_model import models
-from .base import BaseCheck
-from .base import CheckLevel
-from .geo_query import distance
-from .geo_query import length
-from .geo_query import transform
 from dataclasses import dataclass
-from geoalchemy2 import functions as geo_func
-from sqlalchemy import func
-from sqlalchemy import text
-from sqlalchemy.orm import aliased
-from sqlalchemy.orm import Query
-from sqlalchemy.orm import Session
-from typing import List
-from typing import NamedTuple
+from typing import List, NamedTuple
+
+from sqlalchemy import func, text
+from sqlalchemy.orm import aliased, Query, Session
+from threedi_schema import constants, models
+
+from .base import BaseCheck, CheckLevel
+from .geo_query import distance, length, transform
 
 
 class CrossSectionLocationCheck(BaseCheck):
@@ -274,8 +267,8 @@ class LinestringLocationCheck(BaseCheck):
         end_node = aliased(models.ConnectionNode)
 
         tol = self.max_distance
-        start_point = geo_func.ST_PointN(self.column, 1)
-        end_point = geo_func.ST_PointN(self.column, geo_func.ST_NPoints(self.column))
+        start_point = func.ST_PointN(self.column, 1)
+        end_point = func.ST_PointN(self.column, func.ST_NPoints(self.column))
 
         start_ok = distance(start_point, start_node.the_geom) <= tol
         end_ok = distance(end_point, end_node.the_geom) <= tol
@@ -379,7 +372,7 @@ class PotentialBreachStartEndCheck(BaseCheck):
         linestring = models.Channel.the_geom
         tol = self.min_distance
         breach_point = func.Line_Locate_Point(
-            transform(linestring), transform(geo_func.ST_PointN(self.column, 1))
+            transform(linestring), transform(func.ST_PointN(self.column, 1))
         )
         dist_1 = breach_point * length(linestring)
         dist_2 = (1 - breach_point) * length(linestring)
@@ -409,7 +402,7 @@ class PotentialBreachInterdistanceCheck(BaseCheck):
         # First fetch the position of each potential breach per channel
         def get_position(point, linestring):
             breach_point = func.Line_Locate_Point(
-                transform(linestring), transform(geo_func.ST_PointN(point, 1))
+                transform(linestring), transform(func.ST_PointN(point, 1))
             )
             return (breach_point * length(linestring)).label("position")