+++ threedi_modelchecker/errors.py
@@ -1,8 +1,0 @@
-class MigrationMissingError(Exception):
-    """Raised when 3Di model is missing migrations."""
-
-    pass
-
-
-class UpgradeFailedError(Exception):
-    """Raised when an upgrade() fails"""