+++ threedi_modelchecker/scripts.py
@@ -1,10 +1,9 @@
+import click
+from threedi_schema import ThreediDatabase
+
 from threedi_modelchecker import exporters
 from threedi_modelchecker.checks.base import CheckLevel
 from threedi_modelchecker.model_checks import ThreediModelChecker
-from threedi_modelchecker.schema import ModelSchema
-from threedi_modelchecker.threedi_database import ThreediDatabase
-
-import click
 
 
 @click.group()
@@ -16,7 +15,7 @@ import click
     required=True,
 )
 @click.pass_context
-def threedi_modelchecker(ctx, sqlite):
+def main(ctx, sqlite):
     """Checks the threedi-model for errors / warnings / info messages"""
     ctx.ensure_object(dict)
 
@@ -24,7 +23,7 @@ def threedi_modelchecker(ctx, sqlite):
     ctx.obj["db"] = db
 
 
-@threedi_modelchecker.command()
+@main.command()
 @click.option("-f", "--file", help="Write errors to file, instead of stdout")
 @click.option(
     "-l",
@@ -58,39 +57,5 @@ def check(ctx, file, level):
     click.echo("Finished processing model")
 
 
-@threedi_modelchecker.command()
-@click.option(
-    "-r", "--revision", default="head", help="The schema revision to migrate to"
-)
-@click.option("--backup/--no-backup", default=True)
-@click.option("--set-views/--no-set-views", default=True)
-@click.option(
-    "--upgrade-spatialite-version/--no-upgrade-spatialite-version", default=False
-)
-@click.pass_context
-def migrate(ctx, revision, backup, set_views, upgrade_spatialite_version):
-    """Migrate the threedi model schematisation to the latest version."""
-    schema = ModelSchema(ctx.obj["db"])
-    click.echo("The current schema revision is: %s" % schema.get_version())
-    click.echo("Running alembic upgrade script...")
-    schema.upgrade(
-        revision=revision,
-        backup=backup,
-        set_views=set_views,
-        upgrade_spatialite_version=upgrade_spatialite_version,
-    )
-    click.echo("The migrated schema revision is: %s" % schema.get_version())
-
-
-@threedi_modelchecker.command()
-@click.pass_context
-def index(ctx):
-    """Set the indexes of a threedi model schematisation."""
-    schema = ModelSchema(ctx.obj["db"])
-    click.echo("Recovering indexes...")
-    schema.set_spatial_indexes()
-    click.echo("Done.")
-
-
 if __name__ == "__main__":
-    threedi_modelchecker()
+    main()