+++ CHANGES.rst
@@ -2,10 +2,15 @@ Changelog of threedi-modelchecker
 =================================
 
 
-0.35.3 (unreleased)
--------------------
+1.0.0 (unreleased)
+------------------
 
-- Nothing changed yet.
+- Separate the schema to a separate package: threedi-schema.
+
+- Removed threedi_modelchecker.schema, threedi_database, threedi_model,
+  ThreediDatabase. Import these from threedi-schema.
+
+- Remove simulation templates generation code.
 
 
 0.35.2 (2023-01-18)